+++ R/plot.R
@@ -266,7 +266,8 @@ plot.big_sp_list <- function(x, coeff = 1, ...) {
     foreach(k = seq_along(mods), .combine = "rbind") %do% {
       mod <- mods[[k]]
       loss <- mod$loss.val
-      cbind.data.frame(set = k, alpha = mod$alpha, message = mod$message,
+      cbind.data.frame(set = k, alpha = mod$alpha, pf.num = mod$pf.num,
+                       message = mod$message,
                        loss_index = seq_along(loss), loss = loss)
     }
   }
@@ -274,7 +275,7 @@ plot.big_sp_list <- function(x, coeff = 1, ...) {
   ggplot(info) +
     theme_bigstatsr(size.rel = coeff) +
     geom_point(aes(loss_index, loss, color = as.factor(set))) +
-    facet_wrap(~alpha, labeller = signif) +
+    facet_grid(pf.num ~ alpha, labeller = signif) +
     scale_colour_discrete(guide = FALSE) +
     labs(x = "Index", y = "Loss for each validation set")
 }