+++ tests/testthat/test-spLogReg.R
@@ -132,11 +132,11 @@ if (not_cran) {
                           covar.row = covar[-ind, ])
         auc0 <- AUC(preds4, y[-ind])
 
-        pf <- rep(1, ncol(X)); pf[set] <- 10
+        pf0 <- rep(1, ncol(X)); pf1 <- pf0; pf1[set] <- 10
         expect_warning(
           mod.bigstatsr5 <- big_spLogReg(X, y[ind], ind.train = ind,
                                          covar.train = covar[ind, ],
-                                         alphas = alphas, pf.X = pf,
+                                         alphas = alphas, pf.X = pf1,
                                          ncores = test_cores()),
           "10 variables with low/no variation have been removed.", fixed = TRUE
         )
@@ -144,19 +144,32 @@ if (not_cran) {
                           covar.row = covar[-ind, ])
         expect_lt(AUC(preds5, y[-ind]), auc0)
 
-        pf[set] <- 0
+        pf2 <- pf0; pf2[set] <- 0.01
         expect_warning(
           mod.bigstatsr6 <- big_spLogReg(X, y[ind], ind.train = ind,
                                          covar.train = covar[ind, ],
-                                         alphas = alphas, pf.X = pf,
+                                         alphas = alphas, pf.X = pf2,
                                          ncores = test_cores()),
           "10 variables with low/no variation have been removed.", fixed = TRUE
         )
-        lapply(unlist(mod.bigstatsr6, recursive = FALSE),
-               function(mod) expect_true(all(mod$beta[set2] != 0)))
+        # lapply(unlist(mod.bigstatsr6, recursive = FALSE),
+        #        function(mod) expect_true(all(mod$beta[set2] != 0)))
         preds6 <- predict(mod.bigstatsr6, X, ind.row = (1:N)[-ind],
                           covar.row = covar[-ind, ])
         expect_gt(AUC(preds6, y[-ind]), auc0)
+
+        expect_warning(
+          mod.bigstatsr7 <- big_spLogReg(X, y[ind], ind.train = ind,
+                                         covar.train = covar[ind, ],
+                                         alphas = alphas,
+                                         pf.X = cbind(pf0, pf1, pf2),
+                                         ncores = test_cores()),
+          "10 variables with low/no variation have been removed.", fixed = TRUE
+        )
+        expect_length(mod.bigstatsr7, 6)
+        expect_equal(summary(mod.bigstatsr7)$pf.num, c(3, 3, 1, 1, 2, 2))
+        expect_identical(summary(mod.bigstatsr7, best.only = TRUE),
+                         summary(mod.bigstatsr7)[1, ])
       }
     }
   })